# apps/api/src/modules/admin/admin-roles.controller.ts
import { Body, Controller, Delete, Get, HttpCode, Param, Patch, Post } from "@nestjs/common";
import { ApiTags } from "@nestjs/swagger";
import { creerRoleRequeteSchema, modifierRoleRequeteSchema, roleVueSchema, type RoleVue } from "@pgd/contracts";
import { ApiZodBody, ApiZodResponse } from "../../common/swagger/zod-schema";
import { Roles } from "../../common/decorators/roles.decorator";
import { AdminRolesService } from "./services/admin-roles.service";

@ApiTags("admin")
@Controller("admin/roles")
export class AdminRolesController {
  constructor(private readonly roles: AdminRolesService) {}

  @Roles("ADMIN_PGD")
  @Get()
  async lister(): Promise<RoleVue[]> {
    return this.roles.lister();
  }

  @Roles("ADMIN_PGD")
  @Get(":code")
  @ApiZodResponse(200, roleVueSchema)
  async trouver(@Param("code") code: string): Promise<RoleVue> {
    return this.roles.trouver(code);
  }

  @Roles("ADMIN_PGD")
  @Post()
  @HttpCode(201)
  @ApiZodBody(creerRoleRequeteSchema)
  @ApiZodResponse(201, roleVueSchema)
  async creer(@Body() body: unknown): Promise<RoleVue> {
    const dto = creerRoleRequeteSchema.parse(body);
    return this.roles.creer(dto);
  }

  @Roles("ADMIN_PGD")
  @Patch(":code")
  @ApiZodBody(modifierRoleRequeteSchema)
  @ApiZodResponse(200, roleVueSchema)
  async modifier(@Param("code") code: string, @Body() body: unknown): Promise<RoleVue> {
    const dto = modifierRoleRequeteSchema.parse(body);
    return this.roles.modifier(code, dto);
  }

  @Roles("ADMIN_PGD")
  @Delete(":code")
  async supprimer(@Param("code") code: string): Promise<{ supprime: true }> {
    await this.roles.supprimer(code);
    return { supprime: true };
  }
}