# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/parametres/{circuit}": {
      "get": {
        "operationId": "AdminParametresCalculController_trouver",
        "parameters": [
          {
            "name": "circuit",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "circuit": {
                      "type": "string",
                      "enum": [
                        "DOBB",
                        "DXC",
                        "DF"
                      ]
                    },
                    "tauxTsc": {
                      "type": "number"
                    },
                    "tauxTva": {
                      "type": "number"
                    },
                    "tscActiveDefaut": {
                      "type": "boolean"
                    },
                    "tvaActiveDefaut": {
                      "type": "boolean"
                    },
                    "devise": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "circuit",
                    "tauxTsc",
                    "tauxTva",
                    "tscActiveDefaut",
                    "tvaActiveDefaut",
                    "devise"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      },
      "patch": {
        "operationId": "AdminParametresCalculController_modifier",
        "parameters": [
          {
            "name": "circuit",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tauxTsc": {
                    "type": "number",
                    "minimum": 0
                  },
                  "tauxTva": {
                    "type": "number",
                    "minimum": 0
                  },
                  "tscActiveDefaut": {
                    "type": "boolean"
                  },
                  "tvaActiveDefaut": {
                    "type": "boolean"
                  },
                  "devise": {
                    "type": "string"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "parametre": {
                      "type": "object",
                      "properties": {
                        "circuit": {
                          "type": "string",
                          "enum": [
                            "DOBB",
                            "DXC",
                            "DF"
                          ]
                        },
                        "tauxTsc": {
                          "type": "number"
                        },
                        "tauxTva": {
                          "type": "number"
                        },
                        "tscActiveDefaut": {
                          "type": "boolean"
                        },
                        "tvaActiveDefaut": {
                          "type": "boolean"
                        },
                        "devise": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "circuit",
                        "tauxTsc",
                        "tauxTva",
                        "tscActiveDefaut",
                        "tvaActiveDefaut",
                        "devise"
                      ],
                      "additionalProperties": false
                    },
                    "demandesBrouillonRecalculees": {
                      "type": "integer",
                      "minimum": 0
                    }
                  },
                  "required": [
                    "parametre",
                    "demandesBrouillonRecalculees"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/admin/roles": {
      "get": {
        "operationId": "AdminRolesController_lister",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "AdminRolesController_creer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "minLength": 1
                  },
                  "libelle": {
                    "type": "string",
                    "minLength": 1
                  },
                  "groupeAd": {
                    "type": "string",
                    "minLength": 1
                  },
                  "niveau": {
                    "type": "integer"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "METIER",
                      "PIVOT",
                      "SYSTEME"
                    ]
                  },
                  "dansMatrice": {
                    "type": "boolean"
                  },
                  "requiertMfa": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "code",
                  "libelle",
                  "groupeAd",
                  "niveau",
                  "type"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "libelle": {
                      "type": "string"
                    },
                    "groupeAd": {
                      "type": "string"
                    },
                    "niveau": {
                      "type": "number"
                    },
                    "type": {
                      "type": "string",
                      "enum": [
                        "METIER",
                        "PIVOT",
                        "SYSTEME"
                      ]
                    },
                    "dansMatrice": {
                      "type": "boolean"
                    },
                    "requiertMfa": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "code",
                    "libelle",
                    "groupeAd",
                    "niveau",
                    "type",
                    "dansMatrice",
                    "requiertMfa"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/roles/{code}": {
      "get": {
        "operationId": "AdminRolesController_trouver",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "libelle": {
                      "type": "string"
                    },
                    "groupeAd": {
                      "type": "string"
                    },
                    "niveau": {
                      "type": "number"
                    },
                    "type": {
                      "type": "string",
                      "enum": [
                        "METIER",
                        "PIVOT",
                        "SYSTEME"
                      ]
                    },
                    "dansMatrice": {
                      "type": "boolean"
                    },
                    "requiertMfa": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "code",
                    "libelle",
                    "groupeAd",
                    "niveau",
                    "type",
                    "dansMatrice",
                    "requiertMfa"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      },
      "patch": {
        "operationId": "AdminRolesController_modifier",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "libelle": {
                    "type": "string",
                    "minLength": 1
                  },
                  "groupeAd": {
                    "type": "string",
                    "minLength": 1
                  },
                  "niveau": {
                    "type": "integer"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "METIER",
                      "PIVOT",
                      "SYSTEME"
                    ]
                  },
                  "dansMatrice": {
                    "type": "boolean"
                  },
                  "requiertMfa": {
                    "type": "boolean"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "libelle": {
                      "type": "string"
                    },
                    "groupeAd": {
                      "type": "string"
                    },
                    "niveau": {
                      "type": "number"
                    },
                    "type": {
                      "type": "string",
                      "enum": [
                        "METIER",
                        "PIVOT",
                        "SYSTEME"
                      ]
                    },
                    "dansMatrice": {
                      "type": "boolean"
                    },
                    "requiertMfa": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "code",
                    "libelle",
                    "groupeAd",
                    "niveau",
                    "type",
                    "dansMatrice",
                    "requiertMfa"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "operationId": "AdminRolesController_supprimer",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/admin/parametres-globaux/{cle}": {
      "get": {
        "operationId": "AdminParametresGlobauxController_trouver",
        "parameters": [
          {
            "name": "cle",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "cle": {
                      "type": "string"
                    },
                    "valeur": {},
                    "libelle": {
                      "type": "string",
                      "nullable": true
                    },
                    "modifiableAdmin": {
                      "type": "boolean"
                    },
                    "dateMaj": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "cle",
                    "libelle",
                    "modifiableAdmin",
                    "dateMaj"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      },
      "patch": {
        "operationId": "AdminParametresGlobauxController_modifier",
        "parameters": [
          {
            "name": "cle",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "valeur": {}
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "cle": {
                      "type": "string"
                    },
                    "valeur": {},
                    "libelle": {
                      "type": "string",
                      "nullable": true
                    },
                    "modifiableAdmin": {
                      "type": "boolean"
                    },
                    "dateMaj": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "cle",
                    "libelle",
                    "modifiableAdmin",
                    "dateMaj"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "PGD — API",
    "description": "Plateforme de Gestion des Dégrèvements — Orange Côte d'Ivoire",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
...
}